refactor: add cast to make clear that unsigned subtraction is safe
In `cmp_alpha()` we were doing this:

    return a_len - b_len;

with two unsigned (`size_t`) values:

The intent was that, when two strings share a prefix (eg. "foo" and
"foobar"), the shorter string should win (ie. come first when ordered
alphabetically). However, because we were using signed values, we
_could_ have actually wrapped around and produced the wrong order in
this case.

In practice, we failed benignly and produced the right result anyway.
Take our "foo" vs "foobar" example, for instance:

- (size_t)3 - (size_t)6 would wrap to a huge unsigned value
  (0xfffffffffffffffd)
- Function return type is `int`, so we'd take the lower 32 bits and
  return 0xfffffffd, which is -3 (ie. two's complement signed
  representation), and gives us a "correct" (negative) answer.

So, it works, but only accidentally. Cast to `long` (signed) to make
clear that we expect small values here. (Even though, in practice, you'd
need an absurdly large length difference of somewhere north of 2 billion
to get a wrongly signed number after truncating to `int`, and we'll
never see that.)

Add a regression test just for completeness, even though, as stated,
we'd never expect this bug to manifest in the real world.

Should have no measurable performance impact; benchmarks here pick up
only noise.

Summary of cpu time and (wall time):

                        best    avg      sd     +/-     p    (best)    (avg)      (sd)     +/-     p
         pathological 0.48863 0.51957 0.07222 [+0.4%] 0.05 (0.48902) (0.52030) (0.07362) [+0.3%]
            command-t 0.29843 0.31715 0.07387 [+1.2%] 0.01 (0.29862) (0.31785) (0.07332) [+1.1%] 0.025
    chromium (subset) 1.20508 1.25565 0.09316 [+0.1%]      (0.28518) (0.30535) (0.06915) [+2.1%]  0.01
     chromium (whole) 0.95083 0.96763 0.05960 [-0.0%]      (0.11101) (0.11778) (0.02655) [-0.0%]
           big (400k) 1.46561 1.49192 0.06380 [-0.2%]      (0.17145) (0.17909) (0.02869) [-1.9%]
                total 4.49241 4.55192 0.13220 [+0.1%]      (1.41101) (1.44037) (0.08538) [+0.6%]

+++ lua/wincent/commandt/lib/matcher.c
@@ -277,7 +277,7 @@ static int cmp_alpha(const void *a, const void *b) {
     size_t b_len = b_str->length;
     int order = strncmp(a_ptr, b_ptr, b_len);
     if (order == 0) {
-        return a_len - b_len; // Shorter string wins.
+        return (long)a_len - (long)b_len; // Shorter string wins.
     } else {
         return order;
     }
